# contracts/console-api.openapi.yaml
openapi: 3.0.3
info:
  title: Keel 控制台 API
  version: 1.0.0-draft
  description: |
    keel-console 九个页面消费的全部接口，由 keel-server 实现。

    这份文件是前端、mock 数据、后端实现三方的唯一事实来源：
    - 前端只调这里定义的接口，不直接调 Langfuse / 薄网关 / Prometheus
    - mock 数据由这份文件生成，接口切真时前端只改 baseURL
    - 后端 P1 按这份文件实现，字段不一致以这里为准

    和 keel/v1 五份契约的关系：那五份是智能体侧协议（/v1/invoke、审计事件、
    追踪属性、manifest、评测集），这份是控制面 REST API，两者独立演进。
    引用到的枚举值必须和 contracts/ 下对应 schema 保持一致。
servers:
  - url: https://keel.example.com/api/v1
    description: prod
  - url: http://localhost:5173/api/v1
    description: 本地 mock（MSW 拦截）
security:
  - bearerAuth: []
...
paths:
...
  /insight/overview:
    get:
      tags: [overview]
      operationId: getOverview
      summary: 总览页全部数据
      description: 一次返回页面所需的全部内容，前端不做二次聚合。
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: range
          in: query
          schema: { type: string, enum: [24h, 7d, 30d], default: 24h }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Overview" }

  # ---------------------------------------------------------------- 智能体
...
  /insight/services:
    get:
      tags: [services]
      operationId: getSharedServices
      summary: 共享服务页全部数据
      description: 服务健康来自 K8s 探测，指标来自 Prometheus，知识库来自 rag-forge。调用方和模型成本按环境过滤。知识库没有环境字段，不按环境拆开。
      parameters:
        - $ref: "#/components/parameters/Env"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SharedServices" }

  # ---------------------------------------------------------------- 链路追踪
  /insight/traces:
    get:
      tags: [traces]
      operationId: listTraces
      summary: trace 列表
      description: |
        原型里这一块是写死的五条「最近调用」，正式实现必须是真列表：
        可按智能体筛选、可分页、可按状态过滤。
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: agent
          in: query
          description: 参与该 trace 的任一智能体，缺省为全部
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/NodeStatus" }
        - name: minDurationMs
          in: query
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/TraceSummary" }
  /insight/traces/{traceId}:
    parameters:
      - name: traceId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [traces]
      operationId: getTrace
      summary: trace 详情
      description: |
        keel-server 调 Langfuse Observations API v2 取全部节点后组装，
        再叠加 keel-audit 的审计事件和 keel-server 的审批记录。
        协作图、泳道时间线、调用树三种视图共用这一份 nodes 数据，前端不再二次请求。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/TraceDetail" }
        "404":
          $ref: "#/components/responses/Error"

  # ---------------------------------------------------------------- 评测中心
...
  /audit/events:
    get:
      tags: [audit]
      operationId: listAuditEvents
      summary: 审计事件查询
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: agent
          in: query
          schema: { type: string }
        - name: risk
          in: query
          schema: { $ref: "#/components/schemas/Risk" }
        - name: action
          in: query
          schema: { $ref: "#/components/schemas/AuditAction" }
        - name: decision
          in: query
          schema: { $ref: "#/components/schemas/AuditDecision" }
        - name: traceId
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/AuditEvent" }
  /audit/verify:
    post:
      tags: [audit]
      operationId: verifyAuditChain
      summary: 手动触发哈希链校验
      parameters:
        - name: agent
          in: query
          schema: { type: string, description: 缺省校验全部智能体 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          checked: { type: integer, description: 校验条数 }
                          intact: { type: boolean }
                          brokenAt:
                            type: string
                            nullable: true
                            description: 断链处的 eventId
                          elapsedMs: { type: integer }
  /audit/exports:
    post:
      tags: [audit]
      operationId: requestAuditExport
      summary: 申请导出（自动建审批单，批准后才生成脱敏文件）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filter]
              properties:
                filter:
                  type: object
                  additionalProperties: true
                  description: 与 /audit/events 相同的筛选条件
                reason: { type: string }
      responses:
        "202":
          description: 已建审批单
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          exportId: { type: string }
                          approvalId: { type: string }

  # ---------------------------------------------------------------- 模型网关
  /insight/costs:
    get:
      tags: [models]
      operationId: getModelGateway
      summary: 模型网关页全部数据
      description: 模型与花费来自薄网关（人民币），单次调用明细在链路追踪页看（来自 Langfuse）。
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: range
          in: query
          schema: { type: string, enum: [24h, 7d, 30d], default: 24h }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ModelGateway" }
  /insight/costs/keys/{alias}/budget:
    post:
      tags: [models]
      operationId: updateModelBudget
      summary: 调整虚拟 Key 的人民币日预算
      description: 保存后写到薄网关这把虚拟 Key，并记一条 config.change 审计。日预算与控制台滑杆一致，¥10 到 ¥200，步长 ¥5。
      parameters:
        - name: alias
          in: path
          required: true
          schema: { type: string, example: askdb-dev }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dailyBudgetCny]
              properties:
                dailyBudgetCny: { type: number, format: float, minimum: 10, maximum: 200 }
      responses:
        "200":
          description: 已写到薄网关
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          alias: { type: string }
                          dailyBudgetCny: { type: number, format: float }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: auth-gateway 签发，aud=keel-api
  parameters:
    Env:
      name: env
      in: query
      schema: { type: string, enum: [all, dev, test, staging, prod], default: all }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, enum: [10, 20, 50, 100], default: 10 }
...
  responses:
    Error:
      description: 统一错误体
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Envelope:
      type: object
      required: [code, traceId]
      properties:
        code:
          type: string
          description: 成功固定为 OK，失败为 contracts/error-codes.yaml 里的错误码
          example: OK
        message: { type: string, nullable: true }
        traceId: { type: string }
        data: {}
    ApiError:
      type: object
      required: [code, message, traceId, retryable]
      properties:
        code: { type: string, example: TOOL_NOT_GRANTED }
        message: { type: string }
        traceId: { type: string }
        retryable: { type: boolean }
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
    EnvName:
      type: string
      enum: [dev, test, staging, prod]
    Risk:
      type: string
      enum: [LOW, MID, HIGH]
    AgentStatus:
      type: string
      description: |
        DRAFT 表示 keel new 预占了名称但还没 register。
        原型里的「门禁未过」不是独立状态，由 status=REGISTERED 且 gatePassed=false 推导。
      enum: [DRAFT, REGISTERED, ONLINE, DEGRADED, OFFLINE, RETIRED]
    NodeStatus:
      type: string
      description: 三档口径，来自 askdb trace.py。degraded 归入 fallback，不单列。
      enum: [ok, fallback, failed]
    ObservationType:
      type: string
      description: 与 langfuse.observation.type 一致
      enum: [agent, generation, tool, retriever, guardrail, event]
...
    AuditAction:
      type: string
      description: 与 contracts/audit-event.schema.json 的 action 枚举完全一致，两边必须同时改。
      enum:
        [
          invoke,
          tool.call,
          sql.execute,
          approval,
          config.change,
          data.export,
          run.suspend,
          run.resume,
          agent.register,
          release.gate,
        ]
    AuditDecision:
      type: string
      enum: [allowed, denied, pending, approved, rejected]
...
    Overview:
      type: object
      properties:
        kpi: { $ref: "#/components/schemas/OverviewKpi" }
        agents:
          type: array
          items: { $ref: "#/components/schemas/AgentSummary" }
        alerts:
          type: array
          items: { $ref: "#/components/schemas/Alert" }
        costByAgent:
          type: array
          items:
            type: object
            properties:
              agent: { type: string }
              costCny: { type: number, format: float }
              dailyBudgetCny: { type: number, format: float }
    OverviewKpi:
      type: object
      properties:
        onlineAgents: { type: integer, example: 7 }
        totalAgents: { type: integer, example: 10 }
        bizCount: { type: integer, example: 5 }
        devCount: { type: integer, example: 5 }
        calls: { type: integer, example: 13195 }
        callsTrendPct: { type: number, format: float, example: 8.2 }
        modelCostCny: { type: number, format: float, example: 101.10 }
        avgScore: { type: number, format: float, example: 0.86 }
        gateThreshold: { type: number, format: float, example: 0.85 }
        pendingApprovals: { type: integer, example: 3 }
    Alert:
      type: object
      properties:
        at: { type: string, format: date-time }
        level: { type: string, enum: [INFO, WARN, ERROR] }
        agent: { type: string, nullable: true }
        kind:
          type: string
          description: 对账 finding 或运行告警
          enum:
            [
              OFFLINE,
              UNREGISTERED,
              ZOMBIE,
              RETIRE_INCOMPLETE,
              VERSION_MISMATCH,
              BUDGET_WARN,
              GATE_FAILED,
              PROMPT_DRIFT,
            ]
        text: { type: string }
        link: { type: string, nullable: true }

    # ------------------------------------------------------------ 智能体
    AgentSummary:
      type: object
      properties:
        name: { type: string, example: offshore-wind }
        displayName: { type: string, example: 风机维检 }
        category: { type: string, enum: [biz, dev] }
        language: { type: string, example: Python }
        runtime: { type: string, enum: [code, dify] }
        template: { type: string, example: graph-agent }
        env: { $ref: "#/components/schemas/EnvName" }
        version: { type: string, example: v0.7.2 }
        status: { $ref: "#/components/schemas/AgentStatus" }
        statusNote:
          type: string
          nullable: true
          example: 0/1 就绪 · CrashLoop
        gatePassed: { type: boolean, nullable: true }
        ownerOrg: { type: string }
        ownerUser: { type: string }
        calls24h: { type: integer }
        callsTotal: { type: integer }
        p95Seconds: { type: number, format: float }
        costCny: { type: number, format: float }
        dailyBudgetCny: { type: number, format: float }
        score: { type: number, format: float, nullable: true }
        instances: { type: string, example: "2/2", description: 就绪/期望；Dify 为探活结果 }
        multiAgent: { type: boolean }
        delegateCount:
          type: integer
          description: spec.delegates 的个数，列表卡片用来显示「编排 N 个」
...
    SharedServices:
      type: object
      properties:
        services:
          type: array
          items:
            type: object
            properties:
              name: { type: string, example: rag-forge }
              role: { type: string, example: 知识检索 }
              instances: { type: string, example: "2/2 · worker 1" }
              p95: { type: string, nullable: true, example: 0.78s }
              errorRate: { type: string, example: 0.6%（429） }
              status: { type: string, enum: [ONLINE, DEGRADED, OFFLINE] }
        ragforge:
          type: object
          properties:
            kpi:
              type: object
              properties:
                searches24h: { type: integer }
                searchTrendPct: { type: number, format: float }
                p95Seconds: { type: number, format: float }
                p50Seconds: { type: number, format: float }
                throttleRate: { type: number, format: float }
                kbCount: { type: integer }
                staleKbCount: { type: integer }
                modelCostCny: { type: number, format: float }
            stageLatency:
              type: array
              description: 分段耗时。有直方图分位时用 p50Ms；现网只有 sum/count 时用 meanMs。
              items:
                type: object
                properties:
                  stage: { type: string, enum: [rewrite, vector, keyword, rerank, other] }
                  p50Ms: { type: integer }
                  meanMs: { type: integer }
                  basis: { type: string, enum: [p50, mean] }
            callers:
              type: array
              items:
                type: object
                properties:
                  agent: { type: string }
                  calls: { type: integer }
            knowledgeBases:
              type: array
              items:
                type: object
                properties:
                  kb: { type: string }
                  owner: { type: string }
                  documents: { type: integer }
                  updatedAt: { type: string, format: date-time }
                  searches24h: { type: integer }
                  zeroHitRate: { type: number, format: float }
                  recallAt5: { type: number, format: float }
                  stale: { type: boolean, description: 超过 30 天未更新 }

    # ------------------------------------------------------------ 链路追踪
    TraceSummary:
      type: object
      properties:
        traceId: { type: string, example: tr_8a31f0c2 }
        question: { type: string }
        startedAt: { type: string, format: date-time }
        sessionId: { type: string }
        turn: { type: integer }
        userId: { type: string }
        userRole: { type: string, nullable: true }
        env: { $ref: "#/components/schemas/EnvName" }
        rootAgent: { type: string }
        agents:
          type: array
          items: { type: string }
        multiAgent: { type: boolean }
        durationMs: { type: integer, description: 智能体耗时，不含人工等待 }
        humanWaitMs: { type: integer, nullable: true }
        tokens: { type: integer }
        costCny: { type: number, format: float }
        status: { $ref: "#/components/schemas/NodeStatus" }
        blocked: { type: boolean, description: 被护栏或准入拦截 }
        cached: { type: boolean }
    TraceDetail:
      type: object
      properties:
        summary: { $ref: "#/components/schemas/TraceSummary" }
        langfuseUrl: { type: string, description: "深链 {langfuse}/project/{id}/traces/{traceId}" }
        agents:
          type: array
          description: 参与方，决定协作图与图例的配色
          items:
            type: object
            properties:
              key: { type: string, example: wind }
              name: { type: string, example: offshore-wind }
              subtitle: { type: string, example: 子智能体 · 跨服务 }
              color: { type: string, example: "#2ec4b6" }
              kind: { type: string, enum: [platform, supervisor, sub, service, human] }
        latencyBreakdown:
          type: array
          description: 关键路径构成，各项之和等于 durationMs
          items:
            type: object
            properties:
              label: { type: string, example: offshore-wind }
              ms: { type: integer }
              agentKey: { type: string }
        costBreakdown:
          type: array
          items:
            type: object
            properties:
              agentKey: { type: string }
              costCny: { type: number, format: float }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/TraceNode" }
        edges:
          type: array
          description: 协作图连线；泳道视图用它画交接点
          items:
            type: object
            properties:
              from: { type: string, description: 节点 id }
              to: { type: string }
              label: { type: string, nullable: true, example: 委派 ① }
              atMs: { type: integer, nullable: true }
              parallel: { type: boolean }
    TraceNode:
      type: object
      properties:
        id: { type: string }
        agentKey: { type: string }
        type: { $ref: "#/components/schemas/ObservationType" }
        name: { type: string, example: rag.fault-manual }
        shortName: { type: string, nullable: true, example: 检索手册 }
        service: { type: string, example: rag-forge }
        startMs: { type: integer, description: 相对 trace 起点 }
        durationMs: { type: integer }
        humanWaitLabel: { type: string, nullable: true, example: 4m12s }
        status: { $ref: "#/components/schemas/NodeStatus" }
        depth: { type: integer }
        criticalPath: { type: boolean }
        aggregated: { type: boolean, description: agent 节点，耗时与成本含子节点 }
        model: { type: string, nullable: true }
        fallbackFrom: { type: string, nullable: true }
        llmKeyAlias: { type: string, nullable: true }
        tokensIn: { type: integer, nullable: true }
        tokensOut: { type: integer, nullable: true }
        tokens: { type: integer, nullable: true }
        costCny: { type: number, format: float, nullable: true }
        inputSummary: { type: string, description: Langfuse observation 的 input，没有时用节点名 }
        outputSummary: { type: string, description: Langfuse observation 的 output，没有时用节点名 }
        auditIds:
          type: array
          items: { type: string }
        approvalId: { type: string, nullable: true }
        gateNote: { type: string, nullable: true, example: 评测：askdb v1.9.0 门禁 0.88 通过 }
        deprecatedTool: { type: boolean }

    # ------------------------------------------------------------ 评测
...
    AuditEvent:
      type: object
      description: 与 contracts/audit-event.schema.json 同构，额外带 hash、prevHash、hashVerified 供详情展示哈希链
      properties:
        eventId: { type: string }
        ts: { type: string, format: date-time }
        agent: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        traceId: { type: string, nullable: true }
        actor:
          type: object
          properties:
            userId: { type: string }
            orgId: { type: string, nullable: true }
            role: { type: string, nullable: true }
        action: { $ref: "#/components/schemas/AuditAction" }
        resource: { type: string }
        risk: { $ref: "#/components/schemas/Risk" }
        decision: { $ref: "#/components/schemas/AuditDecision" }
        approver: { type: string, nullable: true }
        payload:
          type: object
          additionalProperties: true
          description: 只含 manifest captureFields 白名单字段，已脱敏
        inputDigest: { type: string }
        hashVerified: { type: boolean }
        hash: { type: string, description: 本事件 sha256 hex }
        prevHash: { type: string, nullable: true, description: 同智能体链上一条的 hash，链头为空 }

    # ------------------------------------------------------------ 模型网关
    ModelGateway:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              model: { type: string, example: qwen-plus }
              provider: { type: string, example: DashScope }
              role: { type: string, example: 主力 }
              calls: { type: integer }
              p95: { type: string, example: 4.6s, description: 窗口内延迟 P95，没有样本时为空 }
              errorRate:
                type: number
                format: float
                description: 窗口内超时次数 / 调用次数。没有延迟或超时样本时为空
              costCny: { type: number, format: float }
              status: { type: string, enum: [ok, warn, bad] }
              priceConfigured:
                type: boolean
                description: 薄网关是否为该模型配置了单价；为 false 时表示配置漂移
        keys:
          type: array
          items:
            type: object
            properties:
              alias: { type: string, example: offshore-wind-prod }
              agent: { type: string }
              env: { $ref: "#/components/schemas/EnvName" }
              models:
                type: array
                items: { type: string }
              dailyBudgetCny: { type: number, format: float }
              spentCny: { type: number, format: float }
              status: { type: string, enum: [ACTIVE, BLOCKED] }
        routing:
          type: array
          description: 由各智能体 manifest 生成，写入薄网关的虚拟 Key
          items:
            type: object
            properties:
              agent: { type: string }
              default: { type: string }
              fallback:
                type: array
                items: { type: string }
              timeoutSeconds: { type: integer }